+++ packages/terra-framework-docs/src/terra-dev-site/test/pills/filter-pills/common/PillTemplate.jsx
@@ -62,6 +62,7 @@ const PillTemplate = ({
             id={pill.id}
             key={pill.id}
             metaData={{ index }}
+            isPopupHeaderDisabled={false}
           />
         ))}
       </FilterPills>